# apps/pairing-server/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DropVoice Pairing & Signaling Server API",
    "description": "公网 HTTPS 配对 + WebRTC 信令服务器。部署在 VPS 上，经 Cloudflare 回源到 Caddy:4443，再反向代理到 axum 127.0.0.1:38424。\n\n## 职责边界\n本服务只做**设备注册 + WebRTC 信令转发**，绝不中继文本数据。\n手机扫码后发起 offer，服务器通过 SSE 转发给桌面，桌面回填 answer；\nSDP 交换完成后，手机↔桌面 WebRTC DataChannel P2P 直连，服务器退出。\n\n## 认证\n- `POST /api/devices`：新建设备无需认证；已存在设备须携带 Bearer（当前 token）方可复用/轮换，否则 401——token 绝不复述给无凭据方。\n- `PUT /api/devices/{id}/status`：Bearer token（注册时颁发的 `pairing_token`）。\n- `POST /api/devices/{id}/webrtc/offer`：body 带 code（首次配对）或 token（重连）。\n- `POST /api/devices/{id}/webrtc/subscribe`：Bearer token → 一次性 SSE 票据。\n- `GET /api/devices/{id}/webrtc/events`（SSE）：query `?ticket=`（一次性票据）。\n- `POST /api/devices/{id}/webrtc/answer`：Bearer token。\n- `GET /api/devices/{id}/webrtc/answer/{session_id}`：session_id 自身即为凭据。",
    "contact": {
      "name": "DropVoice"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "https://ps.dropvoice.online",
      "description": "生产环境（Cloudflare → Caddy:4443 → axum:38424）"
    }
  ],
  "paths": {
...
    "/api/devices/{device_id}/webrtc/offer": {
      "post": {
        "tags": [
          "webrtc-signaling"
        ],
        "summary": "手机发起 WebRTC offer（配对/重连）",
        "description": "手机扫码后发起。body 带 code（首次配对）或 token（重连）+ sdp。\n服务器不校验 code/token（§3.2），只验证 device 存在，建 session（60s TTL），\n通过 SSE 推送给桌面。credential 字段透传给桌面校验。\n每 device 并发会话上限 10（§4.6），超限 429。SDP 上限 64KB。",
        "operationId": "createOffer",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "description": "目标桌面设备 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferRequest"
              },
              "examples": {
                "first_pair": {
                  "summary": "首次配对",
                  "value": {
                    "code": "123456",
                    "sdp": "v=0\r\n..."
                  }
                },
                "reconnect": {
                  "summary": "重连",
                  "value": {
                    "sdp": "v=0\r\n...",
                    "token": "dvct_xxx"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "session 创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferResponse"
                },
                "example": {
                  "device": {
                    "id": "550e8400-...",
                    "name": "My PC"
                  },
                  "session_id": "9f1c-..."
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "413": {
            "description": "SDP 载荷超过 64KB"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/DeviceOffline"
          }
        }
      }
    },
    "/api/devices/{device_id}/webrtc/answer/{session_id}": {
      "get": {
        "tags": [
          "webrtc-signaling"
        ],
        "summary": "手机长轮询 answer（hold 30s）",
        "description": "手机 POST offer 后长轮询此端点。\n- answer 就绪 → 200 立即返回（accepted/rejected）\n- 30s 内无 answer → 204（桌面离线或 SSE 未就绪）\n- session 过期/不存在 → 404\nsession_id 是手机刚 POST offer 拿到的（高熵），无需额外认证。\n全局并发等待有上限，超限返回 429。",
        "operationId": "pollAnswer",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "description": "目标桌面设备 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "session_id",
            "in": "path",
            "description": "POST offer 返回的会话 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "answer 就绪",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResponse"
                }
              }
            }
          },
          "204": {
            "description": "answer 尚未就绪（30s 超时），继续轮询"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/devices/{device_id}/webrtc/subscribe": {
      "post": {
        "tags": [
          "webrtc-signaling"
        ],
        "summary": "桌面换取 SSE 一次性订阅票据",
        "description": "桌面持 Bearer pairing_token 调用，换取 60 秒单次使用的订阅票据，\n随后以 `GET .../webrtc/events?ticket=` 建立 SSE 长连接。\n票据与 device_id 绑定，兑换即失效。",
        "operationId": "subscribeTicket",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "description": "桌面自身设备 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "票据签发",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/devices/{device_id}/webrtc/events": {
      "get": {
        "tags": [
          "webrtc-signaling"
        ],
        "summary": "桌面 SSE 长连接（一次性票据认证）",
        "description": "桌面维持 1 条 SSE 长连接，接收手机的 incoming offer。\n认证用 query 参数 `?ticket=`（由 `POST .../webrtc/subscribe` 以 Bearer 换取，\n60s 单次使用——长效 pairing_token 不进 URL，避免进入访问日志）。\n每 15s 发具名 `ping` 事件（防 CF 125s 超时 + 连接活性监测）。\noffer 事件 data 格式：`{session_id, credential, sdp}`。",
        "operationId": "subscribeEvents",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "description": "桌面自身设备 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "ticket",
            "in": "query",
            "description": "一次性订阅票据（Bearer 换取）",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE 事件流",
            "content": {
              "text/event-stream": {
                "example": "event: offer\ndata: {\"session_id\":\"...\",\"credential\":\"123456\",\"sdp\":\"v=0...\"}\n\nevent: ping\ndata: {\"ts\":1723000000}\n\n"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/devices/{device_id}/webrtc/answer": {
      "post": {
        "tags": [
          "webrtc-signaling"
        ],
        "summary": "桌面回填 answer",
        "description": "桌面收到 SSE offer 后，校验 credential（本地内存比对 code/token），\n生成 answer SDP 回填。accepted 时附 sdp，rejected 时附 reason（如 invalid_credential）。\n回填后唤醒手机的长轮询。",
        "operationId": "submitAnswer",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "description": "桌面自身设备 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              },
              "examples": {
                "accepted": {
                  "summary": "接受",
                  "value": {
                    "sdp": "v=0\r\n...",
                    "session_id": "9f1c...",
                    "status": "accepted"
                  }
                },
                "rejected": {
                  "summary": "拒绝（credential 错误）",
                  "value": {
                    "reason": "invalid_credential",
                    "session_id": "9f1c...",
                    "status": "rejected"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "回填成功"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnswerResponse": {
        "type": "object",
        "description": "`GET /webrtc/answer/{session_id}` 响应体。",
        "required": [
          "status"
        ],
        "properties": {
          "sdp": {
            "type": [
              "string",
              "null"
            ],
            "description": "accepted 时有值。"
          },
          "status": {
            "$ref": "#/components/schemas/AnswerDecision"
          },
          "reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "rejected 时的原因。"
          }
        }
      },
...
      "OfferRequest": {
        "type": "object",
        "description": "`POST /webrtc/offer` 请求体。code（首次配对）或 token（重连）二选一。",
        "required": [
          "sdp"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ],
            "description": "6 位配对码（首次配对）。"
          },
          "token": {
            "type": [
              "string",
              "null"
            ],
            "description": "`dvct_` 连接令牌（重连）。"
          },
          "sdp": {
            "type": "string",
            "description": "手机生成的 offer SDP。"
          }
        }
      },
      "OfferResponse": {
        "type": "object",
        "description": "`POST /webrtc/offer` 响应体。",
        "required": [
          "session_id",
          "device"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          },
          "device": {
            "$ref": "#/components/schemas/DeviceSummary"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数无效",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            },
            "example": {
              "error": {
                "code": "INVALID_REQUEST",
                "details": {
                  "field": "device_id"
                },
                "message": "The request payload is invalid"
              }
            }
          }
        }
      },
      "DeviceOffline": {
        "description": "桌面无活跃 SSE 订阅（离线），offer 被快速拒绝",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            },
            "example": {
              "error": {
                "code": "DEVICE_OFFLINE",
                "message": "device not reachable / no active subscriber"
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "设备不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            },
            "example": {
              "error": {
                "code": "DEVICE_NOT_FOUND",
                "message": "The device does not exist"
              }
            }
          }
        }
      },
      "RateLimited": {
        "description": "速率超限（RATE_LIMIT 1 req/IP/秒）",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            },
            "example": {
              "error": {
                "code": "RATE_LIMITED",
                "message": "Too many requests"
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "注册时颁发的 64 字符 `pairing_token`。"
      }
    }
  },
...
}